+++ apps/pie-storybook/stories/pie-thumbnail.stories.ts
@@ -1,9 +1,15 @@
 import { html } from 'lit';
+import { ifDefined } from 'lit/directives/if-defined.js';
 import { type Meta } from '@storybook/web-components';
 
 import '@justeattakeaway/pie-thumbnail';
 import {
-    type ThumbnailProps, defaultProps, variants, backgroundColors,
+    type ThumbnailProps,
+    defaultProps,
+    variants,
+    backgroundColors,
+    sizes,
+    aspectRatios,
 } from '@justeattakeaway/pie-thumbnail';
 
 import { createStory, type TemplateFunction } from '../utilities';
@@ -12,8 +18,8 @@ type ThumbnailStoryMeta = Meta<ThumbnailProps>;
 
 const defaultArgs: ThumbnailProps = {
     ...defaultProps,
-    src: 'https://www.pie.design/assets/img/jet-logo-narrow.svg',
-    alt: 'JET logo',
+    src: './static/images/pie-logo.svg',
+    alt: 'The PIE design system logo',
     placeholder: {
         src: 'https://www.pie.design/assets/img/404_narrow.png',
         alt: 'Thumbnail placeholder image',
@@ -30,6 +36,14 @@ const thumbnailStoryMeta: ThumbnailStoryMeta = {
             options: variants,
             defaultValue: {
                 summary: defaultProps.variant,
+            },
+        },
+        size: {
+            description: 'Set the size of the thumbnail.',
+            control: 'select',
+            options: sizes,
+            defaultValue: {
+                summary: defaultArgs.size,
             },
         },
         src: {
@@ -68,8 +82,16 @@ const thumbnailStoryMeta: ThumbnailStoryMeta = {
                 summary: defaultProps.backgroundColor,
             },
         },
+        aspectRatio: {
+            description: 'Sets the aspect-ratio of the thumbnail image.',
+            control: 'select',
+            options: aspectRatios,
+            defaultValue: {
+                summary: defaultProps.aspectRatio,
+            },
+        },
         placeholder: {
-            description: 'If an image is unavailable, the placeholder prop can be used to ensure there is always something visible to users.',
+            description: 'If an image fails to load, the placeholder prop can be used to ensure there is always something visible to users.',
             control: 'object',
             defaultValue: {
                 summary: defaultProps.placeholder,
@@ -89,20 +111,24 @@ export default thumbnailStoryMeta;
 
 const Template: TemplateFunction<ThumbnailProps> = ({
     variant,
+    size,
     src,
     alt,
     disabled,
     hasPadding,
     backgroundColor,
     placeholder,
+    aspectRatio,
 }) => html`
     <pie-thumbnail
-        variant="${variant}"
-        src="${src}"
-        alt="${alt}"
+        variant="${ifDefined(variant)}"
+        size="${ifDefined(size)}"
+        aspectRatio="${ifDefined(aspectRatio)}"
+        src="${ifDefined(src)}"
+        alt="${ifDefined(alt)}"
+        backgroundColor="${ifDefined(backgroundColor)}"
         ?disabled="${disabled}"
         ?hasPadding="${hasPadding}"
-        backgroundColor="${backgroundColor}"
         .placeholder="${placeholder}"
     </pie-thumbnail>`;
 
@@ -112,4 +138,18 @@ export const Default = createThumbnailStory({}, {});
 
 export const Outline = createThumbnailStory({
     variant: 'outline',
+}, {});
+
+export const AspectRatio4By3 = createThumbnailStory({
+    size: 120,
+    src: './static/images/burger-4by3.png',
+    alt: 'Burger King meal with a cheeseburger, fries, onion rings, and Coca-Cola.',
+    aspectRatio: '4by3',
+}, {});
+
+export const AspectRatio16By9 = createThumbnailStory({
+    size: 128,
+    src: './static/images/burger-16by9.png',
+    alt: 'Chef assembling a burger in a kitchen.',
+    aspectRatio: '16by9',
 }, {});