+++ packages/components/pie-radio/test/accessibility/pie-radio.spec.ts
@@ -1,16 +1,10 @@
-import { test, expect } from '@justeattakeaway/pie-webc-testing/src/playwright/webc-fixtures.ts';
-import { PieRadio, type RadioProps } from '../../src/index.ts';
+import { test, expect } from '@justeattakeaway/pie-webc-testing/src/playwright/playwright-fixtures.ts';
+import { BasePage } from '@justeattakeaway/pie-webc-testing/src/helpers/page-object/base-page.ts';
 
 test.describe('PieRadio - Accessibility tests', () => {
-    test('a11y - should test the PieRadio component WCAG compliance', async ({ makeAxeBuilder, mount }) => {
-        await mount(PieRadio, {
-            props: {
-                name: 'option-1',
-            } as RadioProps,
-            slots: {
-                default: 'Label',
-            },
-        });
+    test('a11y - should test the PieRadio component WCAG compliance', async ({ makeAxeBuilder, page }) => {
+        const radioDefaultPage = await new BasePage(page, 'radio--default');
+        await radioDefaultPage.load();
 
         const results = await makeAxeBuilder().analyze();