+++ packages/components/pie-radio/test/helpers/page-object/selectors.ts
@@ -1,0 +1,13 @@
+const radio = {
+    selectors: {
+        container: {
+            dataTestId: 'pie-radio',
+        },
+        input: {
+            dataTestId: 'pie-radio-input',
+        },
+    },
+};
+export {
+    radio,
+};