+++ packages/tools/pie-icons-webc/test/component/icon-webc.spec.js
@@ -1,90 +1,83 @@
 
-import { test, expect } from '@sand4rt/experimental-ct-web';
+import { test, expect } from '@playwright/test';
 import { sizeToValueMap, largeIconSizeDefault, regularIconSizeDefault } from '@justeattakeaway/pie-icons-configs';
-import { IconAlcoholFilled } from '../../icons/IconAlcoholFilled.ts';
-import { IconAlcoholFilledLarge } from '../../icons/IconAlcoholFilledLarge.ts';
+import { BasePage } from '@justeattakeaway/pie-webc-testing/src/helpers/page-object/base-page.ts';
+import { icons } from '../helpers/page-object/selectors.ts';
 
-const iconAlcoholFilledSelector = 'icon-alcohol-filled';
-const iconAlcoholFilledLargeSelector = 'icon-alcohol-filled-large';
-
-test('regular icon should use the default width and height when not provided', async ({ mount, page }) => {
+test('regular icon should use the default width and height when not provided', async ({ page }) => {
     // Arrange
-    await mount(IconAlcoholFilled);
-    await page.waitForSelector(iconAlcoholFilledSelector);
+    const regularIconPage = new BasePage(page, 'icons--alcohol-filled-regular-icon');
+    await regularIconPage.load();
+    await page.waitForSelector(icons.selectors.alcoholFilled.dataTestId);
 
     // Assert
     const exepectedValue = sizeToValueMap[regularIconSizeDefault].toString();
 
-    const svg = await page.locator(`${iconAlcoholFilledSelector} svg`);
+    const svg = page.locator(`${icons.selectors.alcoholFilled.dataTestId} svg`);
     await expect(svg).toHaveAttribute('width', exepectedValue);
     await expect(svg).toHaveAttribute('height', exepectedValue);
 });
 
-test('regular icon should fall back to the default width and height if an invalid value is given', async ({ mount, page }) => {
+test('regular icon should fall back to the default width and height if an invalid value is given', async ({ page }) => {
     // Arrange
-    await mount(IconAlcoholFilled, {
-        props: {
-            size: 'xm',
-        },
-    });
-
-    await page.waitForSelector(iconAlcoholFilledSelector);
+    const regularIconPage = new BasePage(page, 'icons--alcohol-filled-regular-icon');
+    await regularIconPage.load({ size: 'xm' });
+    await page.waitForSelector(icons.selectors.alcoholFilled.dataTestId);
 
     // Assert
     const exepectedValue = sizeToValueMap[regularIconSizeDefault].toString();
 
-    const svg = await page.locator(`${iconAlcoholFilledSelector} svg`);
+    const svg = page.locator(`${icons.selectors.alcoholFilled.dataTestId} svg`);
     await expect(svg).toHaveAttribute('width', exepectedValue);
     await expect(svg).toHaveAttribute('height', exepectedValue);
 });
 
-test('regular should have the default classes applied', async ({ mount, page }) => {
+test('regular should have the default classes applied', async ({ page }) => {
     // Arrange
-    await mount(IconAlcoholFilled);
-    await page.waitForSelector(iconAlcoholFilledSelector);
+    const regularIconPage = new BasePage(page, 'icons--alcohol-filled-regular-icon');
+    await regularIconPage.load();
+    await page.waitForSelector(icons.selectors.alcoholFilled.dataTestId);
 
-    const icon = await page.locator(iconAlcoholFilledSelector);
-
+    // Assert
+    const icon = page.locator(icons.selectors.alcoholFilled.dataTestId);
     await expect(icon).toHaveClass('c-pieIcon c-pieIcon--alcoholFilled');
 });
 
-test('large icon should use the default width and height when not provided', async ({ mount, page }) => {
+test('large icon should use the default width and height when not provided', async ({ page }) => {
     // Arrange
-    await mount(IconAlcoholFilledLarge);
-    await page.waitForSelector(iconAlcoholFilledLargeSelector);
+    const largeIconPage = new BasePage(page, 'icons--alcohol-filled-large-icon');
+    await largeIconPage.load();
+    await page.waitForSelector(icons.selectors.alcoholFilledLarge.dataTestId);
 
     // Assert
     const exepectedValue = largeIconSizeDefault.toString();
 
-    const svg = await page.locator(`${iconAlcoholFilledLargeSelector} svg`);
+    const svg = page.locator(`${icons.selectors.alcoholFilledLarge.dataTestId} svg`);
     await expect(svg).toHaveAttribute('width', exepectedValue);
     await expect(svg).toHaveAttribute('height', exepectedValue);
 });
 
-test('large icon should fall back to the default width and height if an invalid value is given', async ({ mount, page }) => {
+test('large icon should fall back to the default width and height if an invalid value is given', async ({ page }) => {
     // Arrange
-    await mount(IconAlcoholFilledLarge, {
-        props: {
-            size: '55',
-        },
-    });
-
-    await page.waitForSelector(iconAlcoholFilledLargeSelector);
+    const largeIconPage = new BasePage(page, 'icons--alcohol-filled-large-icon');
+    await largeIconPage.load({ size: '55' });
+    await page.waitForSelector(icons.selectors.alcoholFilledLarge.dataTestId);
 
     // Assert
     const exepectedValue = largeIconSizeDefault.toString();
 
-    const svg = await page.locator(`${iconAlcoholFilledLargeSelector} svg`);
+    const svg = page.locator(`${icons.selectors.alcoholFilledLarge.dataTestId} svg`);
     await expect(svg).toHaveAttribute('width', exepectedValue);
     await expect(svg).toHaveAttribute('height', exepectedValue);
 });
 
-test('large should have the default classes applied', async ({ mount, page }) => {
+test('large should have the default classes applied', async ({ page }) => {
     // Arrange
-    await mount(IconAlcoholFilledLarge);
-    await page.waitForSelector(iconAlcoholFilledLargeSelector);
+    const largeIconPage = new BasePage(page, 'icons--alcohol-filled-large-icon');
+    await largeIconPage.load();
+    await page.waitForSelector(icons.selectors.alcoholFilledLarge.dataTestId);
 
-    const icon = await page.locator(iconAlcoholFilledLargeSelector);
-
+    // Assert
+    const icon = page.locator(icons.selectors.alcoholFilledLarge.dataTestId);
     await expect(icon).toHaveClass('c-pieIcon c-pieIcon--alcoholFilledLarge');
 });