+++ apps/examples/wc-angular12/CHANGELOG.md
@@ -1,5 +1,12 @@
 # wc-angular12
 
+## 0.5.52
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.6.6
+
 ## 0.5.51
 
 ### Patch Changes