+++ packages/components/pie-toast/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @justeattakeaway/pie-toast
 
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [[`a643b87`](https://github.com/justeattakeaway/pie/commit/a643b876e66d86effc318d0f0bb98b06508ee003)]:
+  - @justeattakeaway/pie-icons-webc@1.3.0
+  - @justeattakeaway/pie-icon-button@1.1.1
+
 ## 0.7.0
 
 ### Minor Changes